docs: add quarto and quartodoc for static website (#38)
`_quarto.yml` includes the base structure of the static website with a navbar and the main site.  The navbar includes 'Overview', 'Get Started', and 'API Referece' sections, which are structured in `_index.qmd`, `_get_started.qmd`, and the quartodoc section of `_quarto.yml` respectively. Currently, the 'Overview' and 'Get Started' sections are left empty, which will be generated in the upcoming releases. The 'API Reference' section will include the docstring in a neat format generated by `quartodoc`.

Resolves: #17