ci(semantic-release): exclude non-conventional commits from `CHANGELOG` (#106)
This pull request updates the `exclude_commit_patterns` in the
`pyproject.toml` file. The previous commits that don't match the
conventional commits prefixes and internal changes that do not
necessarily affect end-user interactions, such as `chore`, `ci`, and
`style`, are excluded from our CHANGELOG and GitHub release note moving
forward. This is not a direct fix, but after this change, it ensures
that only relevant commits are included in the release changelog as a
fair stopgap solution.

Closes #99